requirements.txt
aiogram==3.13.0
aiohttp==3.9.5
beautifulsoup4==4.12.3
python-dotenv==1.0.1
flask==3.0.0
gunicorn==21.2.0
